# perf: avoid overuse of interfaces