--- a/effects.go
+++ b/effects.go
@@ -4,11 +4,13 @@ type ErrFn func() error
 
 func Effect(rs *ReactiveSystem, fn ErrFn) ErrFn {
 	e := &EffectRunner{
-		fn: fn,
+		fn:             fn,
+		baseDependency: baseDependency{},
 		baseSubscriber: baseSubscriber{
-			_flags: fEffect,
+			flags: fEffect,
 		},
 	}
+	sub := e.sub()
 
 	if rs.activeSub != nil {
 		rs.link(e, rs.activeSub)
@@ -18,8 +20,8 @@ func Effect(rs *ReactiveSystem, fn ErrFn) ErrFn {
 	rs.runEffect(e)
 
 	return func() error {
-		rs.startTracking(e)
-		rs.endTracking(e)
+		rs.startTracking(sub)
+		rs.endTracking(sub)
 		return nil
 	}
 }
@@ -27,33 +29,34 @@ func Effect(rs *ReactiveSystem, fn ErrFn) ErrFn {
 func (rs *ReactiveSystem) runEffect(e *EffectRunner) {
 	prevSub := rs.activeSub
 	rs.activeSub = e
-	rs.startTracking(e)
+	sub := e.sub()
+	rs.startTracking(sub)
 	if err := e.fn(); err != nil {
 		if rs.onError != nil {
 			rs.onError(e, err)
 		}
 	}
-	rs.endTracking(e)
+	rs.endTracking(sub)
 	rs.activeSub = prevSub
 }
 
 func (rs *ReactiveSystem) notifyEffect(sub subscriber) bool {
 	e := mustEffect(sub)
-	flags := e.flags()
+	sub2 := sub.sub()
+	flags := sub2.flags
 	if flags&fEffectScope != 0 {
-		flags := e.flags()
 		if flags&fPendingEffect != 0 {
-			rs.processPendingInnerEffects(e, e.flags())
+			rs.processPendingInnerEffects(sub2, flags)
 			return true
 		}
 		return false
 	}
 
 	if flags&fDirty != 0 ||
-		(flags&fPendingComputed != 0 && rs.updateDirtyFlag(e, flags)) {
+		(flags&fPendingComputed != 0 && rs.updateDirtyFlag(sub2, flags)) {
 		rs.runEffect(e)
 	} else {
-		rs.processPendingInnerEffects(e, flags)
+		rs.processPendingInnerEffects(sub2, flags)
 	}
 	return true
 }
@@ -61,13 +64,14 @@ func (rs *ReactiveSystem) notifyEffect(sub subscriber) bool {
 func EffectScope(rs *ReactiveSystem, scopedFn ErrFn) (stopScope ErrFn) {
 	e := &EffectRunner{
 		baseSubscriber: baseSubscriber{
-			_flags: fEffect | fEffectScope,
+			flags: fEffect | fEffectScope,
 		},
 	}
+	sub := e.sub()
 	rs.runEffectScope(e, scopedFn)
 	return func() error {
-		rs.startTracking(e)
-		rs.endTracking(e)
+		rs.startTracking(sub)
+		rs.endTracking(sub)
 		return nil
 	}
 }
@@ -78,12 +82,21 @@ type EffectRunner struct {
 	fn ErrFn
 }
 
+func (s *EffectRunner) dep() *baseDependency {
+	return &s.baseDependency
+}
+
+func (s *EffectRunner) sub() *baseSubscriber {
+	return &s.baseSubscriber
+}
+
 func (e *EffectRunner) isSignalAware() {}
 
 func (rs *ReactiveSystem) runEffectScope(e *EffectRunner, scopedFn ErrFn) {
 	prevSub := rs.activeSub
 	rs.activeScope = e
-	rs.startTracking(e)
+	sub := e.sub()
+	rs.startTracking(sub)
 
 	if err := scopedFn(); err != nil {
 		if rs.onError != nil {
@@ -92,7 +105,7 @@ func (rs *ReactiveSystem) runEffectScope(e *EffectRunner, scopedFn ErrFn) {
 	}
 
 	rs.activeScope = prevSub
-	rs.endTracking(e)
+	rs.endTracking(sub)
 }
 
 // Ensures all pending internal effects for the given subscriber are processed.
@@ -104,21 +117,18 @@ func (rs *ReactiveSystem) runEffectScope(e *EffectRunner, scopedFn ErrFn) {
 //
 // @param sub - The subscriber which may have pending effects.
 // @param flags - The current flags on the subscriber to check.
-func (rs *ReactiveSystem) processPendingInnerEffects(sub subscriber, flags subscriberFlags) {
+func (rs *ReactiveSystem) processPendingInnerEffects(sub *baseSubscriber, flags subscriberFlags) {
 	if flags&fPendingEffect != 0 {
-		sub.setFlags(flags & ^fPendingEffect)
-		link := sub.deps()
+		sub.flags = flags & ^fPendingEffect
+		link := sub.deps
 		for {
 			dep := link.dep
-			depSub, ok := dep.(dependencyAndSubscriber)
+			depSub, ok := dep.(subscriber)
 			if ok {
-				flags := depSub.flags()
+				depSub2 := depSub.sub()
+				flags := depSub2.flags
 				if flags&fEffect != 0 && flags&fPropagated != 0 {
-					effect, ok := depSub.(*EffectRunner)
-					if !ok {
-						panic("not an effect")
-					}
-					rs.notifyEffect(effect)
+					rs.notifyEffect(depSub)
 				}
 			}
 			link = link.nextDep
@@ -143,7 +153,8 @@ func (rs *ReactiveSystem) processEffectNotifications() {
 			rs.queuedEffectsTail = nil
 		}
 		if !rs.notifyEffect(effect) {
-			effect.setFlags(effect.flags() & ^fNotified)
+			sub := effect.sub()
+			sub.flags = sub.flags & ^fNotified
 		}
 	}
 }