--- a/types.go
+++ b/types.go
@@ -24,66 +24,18 @@ type link struct {
 }
 
 type subscriber interface {
-	flags() subscriberFlags
-	setFlags(subscriberFlags)
-	deps() *link
-	setDeps(*link)
-	depsTail() *link
-	setDepsTail(*link)
-}
-
-type baseSubscriber struct {
-	_flags           subscriberFlags
-	_deps, _depsTail *link
-}
-
-func (s *baseSubscriber) flags() subscriberFlags {
-	return s._flags
-}
-func (s *baseSubscriber) setFlags(flags subscriberFlags) {
-	s._flags = flags
-}
-func (s *baseSubscriber) deps() *link {
-	return s._deps
-}
-func (s *baseSubscriber) setDeps(deps *link) {
-	s._deps = deps
-}
-func (s *baseSubscriber) depsTail() *link {
-	return s._depsTail
-}
-func (s *baseSubscriber) setDepsTail(depsTail *link) {
-	s._depsTail = depsTail
+	sub() *baseSubscriber
 }
 
 type dependency interface {
-	subs() *link
-	setSubs(*link)
-	subsTail() *link
-	setSubsTail(*link)
+	dep() *baseDependency
+}
+
+type baseSubscriber struct {
+	flags          subscriberFlags
+	deps, depsTail *link
 }
 
 type baseDependency struct {
-	_subs, _subsTail *link
-}
-
-func (d *baseDependency) subs() *link {
-	return d._subs
-}
-
-func (d *baseDependency) setSubs(subs *link) {
-	d._subs = subs
-}
-
-func (d *baseDependency) subsTail() *link {
-	return d._subsTail
-}
-
-func (d *baseDependency) setSubsTail(subsTail *link) {
-	d._subsTail = subsTail
-}
-
-type dependencyAndSubscriber interface {
-	dependency
-	subscriber
+	subs, subsTail *link
 }